requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pyaml==17.12.1
PyYAML==3.12
Werkzeug==0.14.1
